requirements.txt
streamlit
google-generativeai
python-dotenv
pathlib
python-docx